sidebar fix
